requirements.txt
Flask
black
psycopg2-binary